XmlReader improvements has been moved to XmlDocumentationExtensions.
