+++ src/DynamoCore/Library/XmlDocumentationExtensions.cs
@@ -3,8 +3,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Text;
-using System.Xml.Linq;
-using System.Xml.XPath;
+using System.Xml;
 using Dynamo.Interfaces;
 using Dynamo.Library;
 
@@ -15,24 +14,61 @@ namespace Dynamo.DSEngine
     /// </summary>
     public static class XmlDocumentationExtensions
     {
+        private static Dictionary<string, MemberDocumentNode> documentNodes =
+                   new Dictionary<string, MemberDocumentNode>();
+
+        #region Public methods
+
+        public static string GetSummary(this FunctionDescriptor member, IPathManager pathManager)
+        {
+            XmlReader xml = null;
+
+            if (member.Assembly != null)
+                xml = DocumentationServices.GetForAssembly(member.Assembly, pathManager);
+
+            return member.GetSummary(xml);
+        }
+
+        public static string GetDescription(this TypedParameter member, IPathManager pathManager)
+        {
+            XmlReader xml = null;
+
+            if (member.Function != null && member.Function.Assembly != null)
+                xml = DocumentationServices.GetForAssembly(member.Function.Assembly, pathManager);
+
+            return member.GetDescription(xml);
+        }
+
+        public static IEnumerable<string> GetSearchTags(this FunctionDescriptor member)
+        {
+            XmlReader xml = null;
+
+            if (member.Assembly != null)
+                xml = DocumentationServices.GetForAssembly(member.Assembly, member.PathManager);
+
+            return member.GetSearchTags(xml);
+        }
+
+        #endregion
+
         #region Overloads specifying XDocument
 
-        public static string GetDescription(this TypedParameter parameter, XDocument xml)
+        public static string GetDescription(this TypedParameter parameter, XmlReader xml)
         {
             if (xml == null) return String.Empty;
 
-            return GetMemberElement(parameter.Function, 
-                String.Format("param[@name='{0}']", parameter.Name), xml).CleanUpDocString();
+            return GetMemberElement(parameter.Function,
+                "description", xml).CleanUpDocString();
         }
 
-        public static string GetSummary(this FunctionDescriptor member, XDocument xml)
+        public static string GetSummary(this FunctionDescriptor member, XmlReader xml)
         {
             if (xml == null) return String.Empty;
 
             return GetMemberElement(member, "summary", xml).CleanUpDocString();
         }
 
-        public static IEnumerable<string> GetSearchTags(this FunctionDescriptor member, XDocument xml)
+        public static IEnumerable<string> GetSearchTags(this FunctionDescriptor member, XmlReader xml)
         {
             if (xml == null) return new List<string>();
 
@@ -66,35 +102,55 @@ namespace Dynamo.DSEngine
             return sb.ToString();
         }
 
-        private static string GetMemberElement(FunctionDescriptor function,
-            string suffix, XDocument xml)
+        private static string GetMemberElement(FunctionDescriptor fDescr, string property, XmlReader xml, string paramName = "")
         {
-            // Construct the entire function descriptor name, including CLR style names
-            string clrMemberName = GetMemberElementName(function);
+            var assemblyName = fDescr.Assembly;
+            var fullyQualifiedName = GetMemberElementName(fDescr);
+            if (!documentNodes.ContainsKey(fullyQualifiedName))
+                LoadDataFromXml(xml);
 
-            // match clr member name
-            var match = xml.XPathEvaluate(
-                String.Format("string(/doc/members/member[@name='{0}']/{1})", clrMemberName, suffix));
+            // Case for operators.
+            if (assemblyName == null)
+                return String.Empty;
 
-            if (match is String && !string.IsNullOrEmpty((string)match))
+            var keyForOverloaded = documentNodes.Keys.
+                        Where(key => key.Contains(fDescr.ClassName + "." + fDescr.FunctionName)).FirstOrDefault();
+
+            switch (property)
             {
-                return match as string;
+                case "summary":
+                    if (documentNodes.ContainsKey(fullyQualifiedName))
+                        return documentNodes[fullyQualifiedName].Summary;
+
+                    // Fallback for overloaded methods.
+                    if (keyForOverloaded != null)
+                        return documentNodes[keyForOverloaded].Summary;
+
+                    return String.Empty;
+
+                case "description":
+                    if (documentNodes.ContainsKey(fullyQualifiedName))
+                        if (documentNodes[fullyQualifiedName].Parameters.ContainsKey(paramName))
+                            return documentNodes[fullyQualifiedName].Parameters[paramName];
+
+                    // Fallback for overloaded methods.
+                    if (keyForOverloaded != null)
+                        if (documentNodes[keyForOverloaded].Parameters.ContainsKey(paramName))
+                            return documentNodes[keyForOverloaded].Parameters[paramName];
+
+                    return String.Empty;
+                case "search":
+                    if (documentNodes.ContainsKey(fullyQualifiedName))
+                        return documentNodes[fullyQualifiedName].SearchTags;
+
+                    // Fallback for overloaded methods.
+                    if (keyForOverloaded != null)
+                        return documentNodes[keyForOverloaded].SearchTags;
+
+                    return String.Empty;
+                default:
+                    return String.Empty;
             }
-
-            // fallback, namespace qualified method name
-            var methodName = function.QualifiedName;
-
-            // match with fallback
-            match = xml.XPathEvaluate(
-                String.Format(
-                    "string(/doc/members/member[contains(@name,'{0}')]/{1})", methodName, suffix));
-
-            if (match is String && !string.IsNullOrEmpty((string)match))
-            {
-                return match as string;
-            }
-
-            return String.Empty;
         }
 
         private static string PrimitiveMap(string s)
@@ -163,6 +219,81 @@ namespace Dynamo.DSEngine
 
             // elements are of the form "M:Namespace.Class.Method"
             return String.Format("{0}:{1}", prefixCode, memberName);
+        }
+
+        private enum XmlTagType
+        {
+            None,
+            Member,
+            Summary,
+            Parameter,
+            SearchTags
+        }
+
+        private static void LoadDataFromXml(XmlReader reader)
+        {
+            if (reader == null)
+                return;
+
+            MemberDocumentNode currentDocNode = new MemberDocumentNode();
+            XmlTagType currentTag = XmlTagType.None;
+            string currentParamName = String.Empty;
+
+            while (reader.Read())
+            {
+                switch (reader.NodeType)
+                {
+                    case XmlNodeType.Element:
+                        switch (reader.Name)
+                        {
+                            case "member":
+                                // Find attribute "name".
+                                if (reader.MoveToAttribute("name"))
+                                {
+                                    currentDocNode = new MemberDocumentNode(reader.Value);
+                                    documentNodes.Add(currentDocNode.FullyQualifiedName, currentDocNode);
+                                }
+                                currentTag = XmlTagType.Member;
+                                break;
+
+                            case "summary":
+                                currentTag = XmlTagType.Summary;
+                                break;
+
+                            case "param":
+                                if (reader.MoveToAttribute("name"))
+                                {
+                                    currentParamName = reader.Value;
+                                }
+                                currentTag = XmlTagType.Parameter;
+                                break;
+
+                            case "search":
+                                currentTag = XmlTagType.SearchTags;
+                                break;
+
+                            default:
+                                currentTag = XmlTagType.None;
+                                break;
+                        }
+                        break;
+                    case XmlNodeType.Text:
+                        switch (currentTag)
+                        {
+                            case XmlTagType.Summary:
+                                currentDocNode.Summary = reader.Value;
+                                break;
+                            case XmlTagType.Parameter:
+                                currentDocNode.Parameters.Add(currentParamName, reader.Value);
+                                break;
+                            case XmlTagType.SearchTags:
+                                currentDocNode.SearchTags = reader.Value;
+                                break;
+                        }
+
+                        break;
+                }
+            }
         }
 
         #endregion
